Added Webpack and made AST interfaces instead of classes
