+++ lib/transformer/src/compiler.ts
@@ -1,0 +1,12 @@
+import { Source } from "./ast";
+
+export interface Program {
+  sources: Source[];
+}
+
+export interface Parser {
+  donelog: Set<string>;
+  seenlog: Set<string>;
+  program: Program;
+  parseFile(sourceText: string, path: string, isEntry: boolean): void;
+}
