+++ lib/transformer/src/examples/index.ts
@@ -1,0 +1,2 @@
+export * from "./functions";
+export * from "./printer";
