+++ lib/transformer/src/parsing.ts
@@ -1,0 +1,194 @@
+
+/** Named token types. */
+export const enum Token {
+
+  // keywords
+  // discarded: ANY, BOOLEAN, NEVER, NUMBER, STRING, SYMBOL, UNDEFINED, LESSTHAN_SLASH
+
+  ABSTRACT,
+  AS,
+  ASYNC,
+  AWAIT,        // ES2017
+  BREAK,        // ES2017
+  CASE,         // ES2017
+  CATCH,        // ES2017
+  CLASS,        // ES2017
+  CONST,        // ES2017
+  CONTINUE,     // ES2017
+  CONSTRUCTOR,
+  DEBUGGER,     // ES2017
+  DECLARE,
+  DEFAULT,      // ES2017
+  DELETE,       // ES2017
+  DO,           // ES2017
+  ELSE,         // ES2017
+  ENUM,         // ES2017 future
+  EXPORT,       // ES2017
+  EXTENDS,      // ES2017
+  FALSE,        // ES
+  FINALLY,      // ES2017
+  FOR,          // ES2017
+  FROM,         // AS possible identifier
+  FUNCTION,     // ES2017
+  GET,
+  IF,           // ES2017
+  IMPLEMENTS,   // ES2017 non-lexical
+  IMPORT,       // ES2017
+  IN,           // ES2017
+  INSTANCEOF,   // ES2017
+  INTERFACE,    // ES2017 non-lexical
+  IS,
+  KEYOF,
+  LET,          // ES2017 non-lexical
+  MODULE,       // AS possible identifier
+  NAMESPACE,    // AS possible identifier
+  NEW,          // ES2017
+  NULL,         // ES
+  OF,
+  PACKAGE,      // ES2017 non-lexical
+  PRIVATE,      // ES2017 non-lexical
+  PROTECTED,    // ES2017 non-lexical
+  PUBLIC,       // ES2017 non-lexical
+  READONLY,
+  RETURN,       // ES2017
+  SET,
+  STATIC,       // ES2017 non-lexical
+  SUPER,        // ES2017
+  SWITCH,       // ES2017
+  THIS,         // ES2017
+  THROW,        // ES2017
+  TRUE,         // ES
+  TRY,          // ES2017
+  TYPE,         // AS possible identifier
+  TYPEOF,       // ES2017
+  VAR,          // ES2017
+  VOID,         // ES2017
+  WHILE,        // ES2017
+  WITH,         // ES2017
+  YIELD,        // ES2017
+
+  // punctuation
+
+  OPENBRACE,
+  CLOSEBRACE,
+  OPENPAREN,
+  CLOSEPAREN,
+  OPENBRACKET,
+  CLOSEBRACKET,
+  DOT,
+  DOT_DOT_DOT,
+  SEMICOLON,
+  COMMA,
+  LESSTHAN,
+  GREATERTHAN,
+  LESSTHAN_EQUALS,
+  GREATERTHAN_EQUALS,
+  EQUALS_EQUALS,
+  EXCLAMATION_EQUALS,
+  EQUALS_EQUALS_EQUALS,
+  EXCLAMATION_EQUALS_EQUALS,
+  EQUALS_GREATERTHAN,
+  PLUS,
+  MINUS,
+  ASTERISK_ASTERISK,
+  ASTERISK,
+  SLASH,
+  PERCENT,
+  PLUS_PLUS,
+  MINUS_MINUS,
+  LESSTHAN_LESSTHAN,
+  GREATERTHAN_GREATERTHAN,
+  GREATERTHAN_GREATERTHAN_GREATERTHAN,
+  AMPERSAND,
+  BAR,
+  CARET,
+  EXCLAMATION,
+  TILDE,
+  AMPERSAND_AMPERSAND,
+  BAR_BAR,
+  QUESTION,
+  COLON,
+  EQUALS,
+  PLUS_EQUALS,
+  MINUS_EQUALS,
+  ASTERISK_EQUALS,
+  ASTERISK_ASTERISK_EQUALS,
+  SLASH_EQUALS,
+  PERCENT_EQUALS,
+  LESSTHAN_LESSTHAN_EQUALS,
+  GREATERTHAN_GREATERTHAN_EQUALS,
+  GREATERTHAN_GREATERTHAN_GREATERTHAN_EQUALS,
+  AMPERSAND_EQUALS,
+  BAR_EQUALS,
+  CARET_EQUALS,
+  AT,
+
+  // literals
+
+  IDENTIFIER,
+  STRINGLITERAL,
+  INTEGERLITERAL,
+  FLOATLITERAL,
+
+  // meta
+
+  INVALID,
+  ENDOFFILE
+}
+
+export function operatorTokenToString(token: Token): string {
+  switch (token) {
+    case Token.DELETE: return "delete";
+    case Token.IN: return "in";
+    case Token.INSTANCEOF: return "instanceof";
+    case Token.NEW: return "new";
+    case Token.TYPEOF: return "typeof";
+    case Token.VOID: return "void";
+    case Token.YIELD: return "yield";
+    case Token.DOT_DOT_DOT: return "...";
+    case Token.COMMA: return ",";
+    case Token.LESSTHAN: return "<";
+    case Token.GREATERTHAN: return ">";
+    case Token.LESSTHAN_EQUALS: return "<=";
+    case Token.GREATERTHAN_EQUALS: return ">=";
+    case Token.EQUALS_EQUALS: return "==";
+    case Token.EXCLAMATION_EQUALS: return "!=";
+    case Token.EQUALS_EQUALS_EQUALS: return "===";
+    case Token.EXCLAMATION_EQUALS_EQUALS: return "!==";
+    case Token.PLUS: return "+";
+    case Token.MINUS: return "-";
+    case Token.ASTERISK_ASTERISK: return "**";
+    case Token.ASTERISK: return "*";
+    case Token.SLASH: return "/";
+    case Token.PERCENT: return "%";
+    case Token.PLUS_PLUS: return "++";
+    case Token.MINUS_MINUS: return "--";
+    case Token.LESSTHAN_LESSTHAN: return "<<";
+    case Token.GREATERTHAN_GREATERTHAN: return ">>";
+    case Token.GREATERTHAN_GREATERTHAN_GREATERTHAN: return ">>>";
+    case Token.AMPERSAND: return "&";
+    case Token.BAR: return "|";
+    case Token.CARET: return "^";
+    case Token.EXCLAMATION: return "!";
+    case Token.TILDE: return "~";
+    case Token.AMPERSAND_AMPERSAND: return "&&";
+    case Token.BAR_BAR: return "||";
+    case Token.EQUALS: return "=";
+    case Token.PLUS_EQUALS: return "+=";
+    case Token.MINUS_EQUALS: return "-=";
+    case Token.ASTERISK_EQUALS: return "*=";
+    case Token.ASTERISK_ASTERISK_EQUALS: return "**=";
+    case Token.SLASH_EQUALS: return "/=";
+    case Token.PERCENT_EQUALS: return "%=";
+    case Token.LESSTHAN_LESSTHAN_EQUALS: return "<<=";
+    case Token.GREATERTHAN_GREATERTHAN_EQUALS: return ">>=";
+    case Token.GREATERTHAN_GREATERTHAN_GREATERTHAN_EQUALS: return ">>>=";
+    case Token.AMPERSAND_EQUALS: return "&=";
+    case Token.BAR_EQUALS: return "|=";
+    case Token.CARET_EQUALS: return "^=";
+    default: {
+      assert(false);
+      return "";
+    }
+  }
+}
