+++ lib/transformer/src/webpack.d.ts
@@ -1,0 +1,4 @@
+declare var DEV: boolean;
+//@ts-ignore
+declare var globalScope: any;
+
