+++ lib/transformer/tests/assembly/funcs.ts
@@ -1,0 +1,5 @@
+function Foo(x: i32): void {}
+
+class FooClass {
+  Foo(x: i32): void {}
+}
