+++ lib/transformer/tests/assembly/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../../../std/assembly.json",
+  "include": [
+    "**/*.ts"
+    ]
+}
