+++ lib/transformer/types/index.d.ts
@@ -1,0 +1,2 @@
+import "./portable";
+import "./common";
