+++ lib/transformer/types/portable/index.d.ts
@@ -1,0 +1,5 @@
+import "../../../../std/types/portable";
+import "../../../src/glue/js/i64";
+import "../../../src/glue/js/float";
+
+
