--- lib/visitor/README.md
@@ -1,6 +1,0 @@
-# AST Transformer
-
-After parsing the files imported by the entry files, the Abstract Syntax Tree (AST) is made.  With the `--transform` argument, the compiler will load any js files provided and if the exports include a Transformer class.
-
-# Transformer class
-
