--- lib/visitor/assembly/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../../std/assembly.json",
-  "include": ["**/*.ts"]
-}
