--- lib/visitor/src/mockTypes.ts
@@ -1,14 +1,0 @@
-import { Source } from "./ast";
-
-//Mock types since imports would add more dependencies than are needed.
-
-export interface Program {
-  sources: Source[];
-}
-
-export interface Parser {
-  donelog: Set<string>;
-  seenlog: Set<string>;
-  program: Program;
-  parseFile(sourceText: string, path: string, isEntry: boolean): void;
-}
