--- lib/visitor/src/webpack.d.ts
@@ -1,4 +1,0 @@
-declare var DEV: boolean;
-declare type FileWriter = (filename: string, contents: Uint8Array | string, baseDir: string) => void
-//@ts-ignore
-declare var globalScope: any;
