--- lib/visitor/tests/assembly/f32.ts
@@ -1,9 +1,0 @@
-const x32: f32 = 1.9;
-
-const y32: f32 = -1.99;
-
-function foo32(f: f32): f32 {
-  return x32 + y32 + f;
-}
-
-foo32(2);
