--- lib/visitor/tests/assembly/f64.ts
@@ -1,9 +1,0 @@
-const x: f64 = 1.9;
-
-const y: f64 = -1.99;
-
-function foo(f: f64): f64 {
-  return x + y + f;
-}
-
-foo(10);
