--- lib/visitor/tests/test.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-set -e
-set -x
-(cd ..; npm run build)
-npm run test
