--- lib/visitor/tests/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../../../std/assembly.json",
-  "include": [
-    "assembly/**/*.ts"
-    ]
-}
