--- lib/visitor/types/portable/index.d.ts
@@ -1,4 +1,0 @@
-import "../../../../std/types/portable";
-import "../../../src/glue/js/i64";
-import "../../../src/glue/js/float";
-
