+++ package.json
@@ -65,6 +65,11 @@
     "lib/loader/index.d.ts",
     "lib/loader/index.js",
     "lib/loader/README.md",
+    "lib/transformer/dist/",
+    "lib/transformer/types/",
+    "lib/transformer/tsconfig",
+    "lib/transformer/src/",
+    "lib/transformer/webpack.base.js",
     "bin/",
     "cli/",
     "dist/",
